staging initial packaging
