+++ requirements.txt
@@ -1,6 +1,9 @@
+biopython==1.85
 black==25.1.0
 click==8.2.1
 mypy_extensions==1.1.0
+numpy==2.2.6
 packaging==25.0
 pathspec==0.12.1
 platformdirs==4.3.8
+setuptools==80.9.0
